Update stock.js to use tmxmoney (#147)
* Update stock.js to use tmxmoney

Quick and dirty modification using axios, cheerio, and string-sanitizer to make the stock plugin work via tmxmoney.com's quote lookup.

* Update stock.js to use tmxmoney

Fixed a typo that did not impact functionality on line 1.

* Update stock.js to use tmxmoney

One final change to default to :US listings in the search results.  This is modifiable via options available in the search box on tmxmoney.  Visit the site, review the options by typing a symbol, and then choose what you want your default to be and modify line 14.