Add support for queuing entire spotify playlists (#186)
This adds support for queuing entire spotify playlists for playback with !playlist. Each song will be searched for with youtube-dl and then queued for playback.

Requires spotify app credentials to use their API.

Should be extensible to other types of playlists by adding more regexps that map to other services.